ref: organizes dirs `include/`&& `src/`
