--- include/scene.h
@@ -1,14 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include "common.h"
-
-
-void InitializeState();
-
-void processEvent(GameObject* object);
-
-GameObject* GetBackground(void);
-
-GameObject* GetObjects(void);
