+++ src/core/game.c
