+++ src/core/player.c
