--- src/scene.c
@@ -1,5 +1,0 @@
-#include "scene.h"
-
-
-GameObject* objects;
-int objectCount;
